requirements.txt
feedparser==6.0.10
Flask==2.3.3
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.8.2